+++ CHANGELOG.md
@@ -7,7 +7,8 @@ All notable changes to `Period` will be documented in this file
 ### Added
 
 - `Duration::create` supports DateInterval spec strings.
-- `Duration::createFromTimer`
+- `Duration::createFromTimeString`
+- `Duration::createFromSeconds`
 
 ### Fixed